Remove ui::TextEditCommand::DELETE_SELECTION.
This CL removes ui::TextEditCommand::DELETE_SELECTION enum value. It is
reduntant since it performs the same operation as a DELETE_FORWARD command.

This is a follow-up to crrev.com/2027133002 and is third in series of CLs to
replace resource ids with a ui::TextEditCommand enum for text editing commands
in views::Textfield.

Link to complete patch - http://crrev.com/2029733003

BUG=586985

Review-Url: https://codereview.chromium.org/2031913002
Cr-Commit-Position: refs/heads/master@{#400936}