Implement distribution of Safe Browsing subresource filtering rules.
Implements the necessary plumbing to distribute read-only mode file handles to
the Safe Browsing subresource filtering ruleset to render processes, to memory
map the contents, and then some more plumbing to disseminate the memory mapped
rules to RenderFrames within the same render process.

For the time being, no ruleset is ever stored in production code, hence no
ruleset is ever distributed to renderers outside of tests. For the sake of
browser tests, the renderer side contains some placeholder logic that treats
the entire ruleset contents as the UTF-8 representation of a string prescribing
a path suffix to filter.

BUG=609747

Review-Url: https://codereview.chromium.org/2077743002
Cr-Commit-Position: refs/heads/master@{#400939}