Refactor ScriptWrappableVisitor.markWrapper(PersistentBase) to be instance method
There is no reason why markWrapper should be static method, and making it
instance method makes it easier to use other visitors (for example verifying
visitor for write barrier checking comes to my mind :)

There is a little fight in me whether to call the method markWrapper or
traceWrappers. I went with traceWrappers in this cl for consistentcy in all
traceWrappers methods. In either case, the same name should be used for
ScopedPersistents too.

LOG=no
BUG=468240

Review-Url: https://codereview.chromium.org/2086553002
Cr-Commit-Position: refs/heads/master@{#400974}