Merge M52: "Immediately reset for vp8 when there's nothing to flush."
If we finish playback of a clip up to end-of-stream, the codec is
already Flush()'d so when we later recieve a Reset(), if we try
to start another Flush() there is no event-pump driving the flush
and we'll hang forever waiting for the flush that never was.

The fix is to instead only flush when there are actually buffers in
the decoder. If there are no buffers, there is nothing to flush.

BUG=619967
TEST=looping video plays

Review-Url: https://codereview.chromium.org/2074983002
Cr-Commit-Position: refs/heads/master@{#400338}
(cherry picked from commit eec6e9013286349a66ace2ec0858c99e96132f46)

Review URL: https://codereview.chromium.org/2084063003 .

Cr-Commit-Position: refs/branch-heads/2743@{#444}
Cr-Branched-From: 2b3ae3b8090361f8af5a611712fc1a5ab2de53cb-refs/heads/master@{#394939}