Implement an observer interface for the site engagement service.
This CL adds the ability for site engagement service clients to register
an observer on the site engagement service. The observer's
OnEngagementIncreased method is called every time engagement is
increased for a URL. The WebContents from which the increase was
triggered and the visible/hidden state of that WebContents is also
passed through to the observer.

BUG=606590

Review-Url: https://codereview.chromium.org/1986033002
Cr-Commit-Position: refs/heads/master@{#401206}