sandwich: Add a timeout of 30 minutes on run.
Before, sandwich was not timing out Chrome run, and could lead to
different source of hanging when recording the trace or tearing
down chrome.

This CL addresses this issue in the SandwichRunner by arming a 30
minutes SIGALARM for each run to fail the task and let the
task_manager carry on on other URLs with the --keep-going flag.

This CL also take the oportunity to let the task_manager to
re-raise ENOSPC errors to avoid corrupting the resuming file.

BUG=582080

Review-Url: https://codereview.chromium.org/2084513004
Cr-Commit-Position: refs/heads/master@{#401247}