headless: Add a dedicated child process entrypoint
Add a dedicated entrypoint for transferring control to the headless
library in a child process (e.g., a renderer). This avoids the need for
the embedder to construct options for the browser, which would get
ignored anyway.

Embedders should now call headless::RunChildProcessIfNeeded() as soon as
possible in their main entrypoint.

BUG=595353

Review-Url: https://codereview.chromium.org/2081423004
Cr-Commit-Position: refs/heads/master@{#401260}