Implement DelegationTracker for tracking delegated permissions to RenderFrameHosts
DelegationTracker maintains a browser-side mapping of which frames have
delegated permissions to other frames. It also allows querying whether a
permission has actually been granted to a frame based on permissions that have
been delegated in the tree of frames and based on the origins of the frames
involved.

For each child RenderFrameHost that permissions are set for, an entry is
added to a map. That entry is removed when the RenderFrameHost is destroyed
or navigated such that permissions will no longer be delegated to it.

A MockRenderFrameHost has been introduced to aid testing.

BUG=614608

Review-Url: https://codereview.chromium.org/2046293002
Cr-Commit-Position: refs/heads/master@{#401478}