Fix audio delay and initial buffer size for very high latency devices.
This improves behavior of the HTML5 audio renderer when the initial
buffer size is so large that it exceeds our previously-static initial
size. Instead we should base this on maxmium of the device period and
the static size.

Another issue that appears is that the delay as calculated by the
OpenSLES output stream is not correct for such large sizes... well
it was probably never correct actually, so this fixes it to use the
same mechanism that MediaPlayer does. Tracking the position of the
stream through time and subtracting from the current position.

BUG=621863
TEST=New unittests, forced --audio-buffer-size=13706 to work well.

Review-Url: https://codereview.chromium.org/2084303002
Cr-Commit-Position: refs/heads/master@{#401486}