SameSite: Correctly set requests' initiator for new tabs.
A user-initiatated navigation in a newly opened tab (via the omnibox, for
example) should be considered SameSite (this is step 1 of Section 2.1
of the spec[1]). This patch adjusts 'FrameLoadRequest' and 'RenderFrameImpl' accordingly.

Now when the browser initiates a navigation on a user's behalf, the frame
type will be set correctly (in 'RenderFrameImpl::NavigateInternal()'), and
the initiator will be updated only if it isn't already set (in
'RenderFrameImpl::willSendRequest()' and 'FrameLoadRequest()').

[1]: https://tools.ietf.org/html/draft-west-first-party-cookies-07#section-2.1

BUG=619603

Review-Url: https://codereview.chromium.org/2080653002
Cr-Commit-Position: refs/heads/master@{#401552}