Implement closed hash map and integer hasher.
ClosedHashMap is an insert-only map implemented as a hash-table with open
addressing (also called closed hashing). The table can contain up to 2^30
distinct keys (or up to 2^32 - 1 on 64bit systems).

On normal operation load factor varies within range (1/4, 1/2]. The real load
factor can be less or equal to 1/4 if rehashing is requested explicitly to
allocate more memory for future insertions.

The table discloses its internal structure in order to allow converting it to
different formats. It consists of 2 vectors, |hash_table| and |entries|. The
|entries| is a vector of pair<KeyType, ValueType>, whereas |hash_table|'s
elements (also called slots) are indexes into the |entries| vector. If, for some
i, hash_table[i] >= entries.size() then the i-th slot is empty. It is guaranteed
that each of the entries is referenced by exactly one table slot.

The Prober is a strategy class used to find a slot for a particular key by
probing the key against a sequence of slots called a probe sequence. Often it
stores one or more hashers - functors used to calculate hash values of the keys
in order to parameterize the probe sequence. The same Prober class can be used
compatibly for performing lookups in the hash map if it's stored in a different
format.

BUG=609747
TBR=caitkp@chromium.org

Review-Url: https://codereview.chromium.org/2084923004
Cr-Commit-Position: refs/heads/master@{#401591}