customtabs: Don't create an extra child process when one is already there.
When a CustomTabActivity is created, the normal initialization sequence
in AsyncInitializationActivity allocates a new child connection. This
asks Android to start a new isolated process. When a spare renderer or a
prerender is used, this is useless and wasteful. Even though this mostly
does not happen on the main thread, child process binding usually takes
>150ms on Android.
Removing this overhead saves memory (a few MB of PSS) and CPU time, and
should speed up Custom Tabs startup, especially on low-end devices.

BUG=607560

Review-Url: https://codereview.chromium.org/2087043003
Cr-Commit-Position: refs/heads/master@{#401609}