sandwich: Implement setup-{prefetch,swr} generating a sandwich_setup.yaml
Before, sandwich had only run commands. But the state could depend
on the state of outside files such as the corpus. This could
lead to non trivial behavior such as when removing an URL, the
directory names of other URLs could change.

This CL fixes that issue by introducing the setup sub-command concept
that prepare the output directory with a sandwich_setup.yaml. Then
the run sub-command just generate the tasks only according to that
file.

Among the changes, the --wpr-archive path is also removed because
now that sandwich support multiple URLs, a single WPR archive
would not scale in size with 100 URLs runs, and also to make sure the
state of the sandwich runs only depends on the state of all the files
stored in the sandwich's output directory.

BUG=582080

Review-Url: https://codereview.chromium.org/2098533002
Cr-Commit-Position: refs/heads/master@{#401646}