[posix] Don't register graceful signal handlers if pipe fails.
This code registers a pipe to listen for graceful shutdown messages,
then registers signal handlers to send graceful shutdown messages.  If
the pipe creation is not successful, the signal handlers cannot work
correctly, so don't register them in that case.

Also add a RAW_CHECK that the setup was complete before a signal was
handled.

BUG=594924

Review-Url: https://codereview.chromium.org/2090163002
Cr-Commit-Position: refs/heads/master@{#401773}