BlueZ + DBus implementations of create/remove service record functions.
This CL adds the DBus and BlueZ implementations of the Create and
Remove service record functions. This adds a basic unit test but
we will need to add more tests to this.

Miao, I've had to change how we handle the service record data structures
since BlueZ will need both the type and size fields from use to correctly
infer the actual data type of the value.

R=mcchou@chromium.org, xiyuan@chromium.org
BUG=619669

Review-Url: https://codereview.chromium.org/2084463002
Cr-Commit-Position: refs/heads/master@{#401782}