Add HTML parser UMA for rewinds and chunk queue statistics.
This patch adds uma for how many tokens we had to rewind the BG parser,
as well as the peak size of the pending chunk queue, in chunk length and
token length. We also log the time it takes to create a chunk/checkpoint,
which can involve lengthy string copies.

This is in preparation for an experiment where the parser's behavior with
the queue will be modified.

BUG=603274

Review-Url: https://codereview.chromium.org/2078723002
Cr-Commit-Position: refs/heads/master@{#401819}