[memory-infra] Add support for queueing memory dump requests in the browser process
This patch changes TracingControllerImpl to queue global memory dump
requests if another dump is in progress (instead of rejecting the
requests). The requests are processed in a simple FIFO order.

To avoid choking the controller, periodic dump requests are skipped if
the queue already contains a (possibly non-periodic) dump request with
the same level of detail.

Rationale: This is necessary for Telemetry tests which combine periodic
and triggered memory dumps.

Note: Support for queueing global memory dump requests in child
processes will be added in a separate patch.

BUG=619935

Review-Url: https://codereview.chromium.org/2067793004
Cr-Commit-Position: refs/heads/master@{#401865}