[Origin Trials] Only flag feature bindings installed after all install methods have returned
This fixes an issue which we haven't run into yet, but will as soon as
an origin trial feature tries to install bindings onto core *and*
modules at the same time. We need to ensure that we only flag the
feature as having had its bindings installed once *all* installation
functions have had a chance to install them.

Review-Url: https://codereview.chromium.org/2096663002
Cr-Commit-Position: refs/heads/master@{#401888}