Fix UrlDataManagerBackendTest when running single-process.
UrlDataMangerBackendTests use the network stack which expects to be run
on the IO thread. This patch configures the TestBrowserThreadBundle to
make the main thread an IO message loop to satisfy that requirement.

This bug is exposed when running content_unittests with the
--single-process-test flag because previous tests leave behind a
URLRequestFilter that includes a DCHECK asserting that it is being
called from a MessageLoopForIO. Otherwise the test gets away with
incorrect usage of the network stack.

BUG=None
R=estade@chromium.org

Review-Url: https://codereview.chromium.org/2091403003
Cr-Commit-Position: refs/heads/master@{#401950}