Weave Packet Generator This CL adds a packet generator that respects the uWeave.BLE protocol.
The packet generator accepts raw bytes as messages to be sent and wraps the message with the uWeave protocol.

The uWeave packets include control packets of request/response/close of a connection and data packets that are the original message chopped up to fixed size packets with uWeave protocol headers.

Design Doc of ProximityAuth uWeave Migration:
https://docs.google.com/a/google.com/document/d/1HJ1fipxsNQwIxmKs-5_6TjroCAcs6qRhAYvOnIr6mQw/edit?usp=sharing

BUG=617238

Review-Url: https://codereview.chromium.org/2031953003
Cr-Commit-Position: refs/heads/master@{#401993}